requirements.txt
asgiref==3.5.2
Django==4.1
django-appconf==1.0.5
django-imagekit==4.1.0
djangorestframework==3.13.1
pilkit==2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
